Реализована возможность добавления записей Не беспокоить

+++ PhoneBook/Controls/AddControl.cs
@@ -1,8 +1,10 @@
-﻿using PhoneBook.Forms;
+﻿using Microsoft.EntityFrameworkCore;
+using PhoneBook.Forms;
 using PhoneBook.Properties;
 using PhoneBook.Types;
 using PhoneBook.UserForms;
 using Syncfusion.WinForms.DataGrid;
+using System.Text.RegularExpressions;
 
 namespace PhoneBook.Controls
 {
@@ -10,6 +12,7 @@ namespace PhoneBook.Controls
     {
         private string _maskNumber = "";
         private NumberPhoneView? currentNumberPhone => gridPhones.DataGrid.SelectedItem as NumberPhoneView;
+        private NotDisturbCollection? currentNotDisturb => gridPhones.DataGrid.SelectedItem as NotDisturbCollection;
 
         public AddControl()
         {
@@ -18,6 +21,11 @@ namespace PhoneBook.Controls
             countryControlAddApartments.CountryChanged += CountryControlAddApartments_CountryChanged;
             cityControlAddApartments.CityChanged += CityControlAddApartments_CityChanged;
             addressControlAddApartments.AddressChanged += AddressControlAddApartments_AddressChanged;
+
+            countryControlAddNotDisturb.CountryChanged += CountryControlAddNotDisturb_CountryChanged;
+            cityControlAddNotDisturb.CityChanged += CityControlAddNotDisturb_CityChanged;
+            addressControlAddNotDisturb.AddressChanged += AddressControlAddNotDisturb_AddressChanged;
+            apartmentControlAddNotDisturb.ApartmentChanged += ApartmentControlAddNotDisturb_ApartmentChanged;
 
             gridPhones.DataGrid.ShowRowHeader = true;
             gridPhones.DataGrid.AutoGeneratingColumn += DataGrid_AutoGeneratingColumn;
@@ -102,22 +110,44 @@ namespace PhoneBook.Controls
                 if (res == DialogResult.OK)
                 {
                     int deletingRowIndex = gridPhones.DataGrid.TableControl.ResolveToRecordIndex(e.DataRow.RowIndex);
-                    using (var db = new ApplicationContext())
+                    if (tabControlAddAdv.SelectedTab.Name == "addApartmentsTab")
                     {
-                        var deleteRow = gridPhones.DataGrid.View.Records[deletingRowIndex].Data as NumberPhoneView;
-
-                        if (deleteRow == null) return;
-
-                        var deleteRecord = db.NumberPhone.Where(n => n.Id == deleteRow.Id).FirstOrDefault();
-                        if (deleteRecord != null)
+                        using (var db = new ApplicationContext())
                         {
-                            db.NumberPhone.Remove(deleteRecord);
-                            db.SaveChanges();
+                            var deleteRow = gridPhones.DataGrid.View.Records[deletingRowIndex].Data as NumberPhoneView;
 
-                            var numberPhones = db.NumberPhoneView
-                                .Where(n => n.AddressId == addressControlAddApartments.GetAddressId())
-                                .ToList();
-                            UpdateData(numberPhones);
+                            if (deleteRow == null) return;
+
+                            var deleteRecord = db.NumberPhone.Where(n => n.Id == deleteRow.Id).FirstOrDefault();
+                            if (deleteRecord != null)
+                            {
+                                db.NumberPhone.Remove(deleteRecord);
+                                db.SaveChanges();
+
+                                var numberPhones = db.NumberPhoneView
+                                    .Where(n => n.AddressId == addressControlAddApartments.GetAddressId())
+                                    .ToList();
+                                UpdateData(numberPhones);
+                            }
+                        }
+                    }
+
+                    if (tabControlAddAdv.SelectedTab.Name == "addNotDisturbTab")
+                    {
+                        using (var db = new ApplicationContext())
+                        {
+                            var deleteRow = gridPhones.DataGrid.View.Records[deletingRowIndex].Data as NotDisturbCollection;
+
+                            if (deleteRow == null) return;
+
+                            var deleteRecord = db.NotDisturb.Where(n => n.Id == deleteRow.Id).FirstOrDefault();
+                            if (deleteRecord != null)
+                            {
+                                db.NotDisturb.Remove(deleteRecord);
+                                db.SaveChanges();
+
+                                UpdateDataNotDisturbBetween();
+                            }
                         }
                     }
                     e.Cancel = true;
@@ -135,7 +165,14 @@ namespace PhoneBook.Controls
         {
             if (tabControlAddAdv.SelectedTab.Name == "addApartmentsTab")
             {
+                UpdateData(new List<NumberPhoneView>());
                 countryControlAddApartments.LoadCountry();
+            }
+
+            if (tabControlAddAdv.SelectedTab.Name == "addNotDisturbTab")
+            {
+                UpdateDataNotDisturb(new List<NotDisturbCollection>());
+                countryControlAddNotDisturb.LoadCountry();
             }
         }
 
@@ -145,6 +182,7 @@ namespace PhoneBook.Controls
             {
                 case "CountryControl":
                     cityControlAddApartments.ClearDataCity();
+                    chkPrivateHouse.Checked = false;
                     chkPrivateHouse.Enabled = false;
                     goto case "CityControl";
                 case "CityControl":
@@ -319,6 +357,205 @@ namespace PhoneBook.Controls
             else
             {
                 MessageBox.Show("Выберите запись для редактирования.", "Уведомление",
+                       MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+        }
+
+        #region Работа с вкладкой Добавить "Не беспокоить"
+        private void UpdateDataNotDisturb(List<NotDisturbCollection> numberPhones)
+        {
+            gridPhones.DataGrid.DataSource = numberPhones;
+            gridPhones.DataGrid.Columns["Locality"].Width = 130;
+            gridPhones.DataGrid.Columns["TypeName"].Width = 90;
+            gridPhones.DataGrid.Columns["StreetName"].AutoSizeColumnsMode = Syncfusion.WinForms.DataGrid.Enums.AutoSizeColumnsMode.LastColumnFill;
+            gridPhones.DataGrid.Columns["House"].Width = 50;
+            gridPhones.DataGrid.Columns["House"].Visible = !chkPrivateHouse.Checked;
+            gridPhones.DataGrid.Columns["Apartment"].Width = 50;
+            gridPhones.DataGrid.Columns["NotCall"].Width = 60;
+            gridPhones.DataGrid.Columns["NotEnter"].Width = 60;
+            gridPhones.DataGrid.Columns["Number"].Width = 120;
+        }
+
+        private void UpdateDataNotDisturbBetween()
+        {
+            if (string.IsNullOrEmpty(addressControlAddNotDisturb.GetTextAddress()))
+                UpdateDataNotDisturb(NotDisturbCollection.GetNotDisturbCollections(cityControlAddNotDisturb.GetCityId()));
+            else
+                UpdateDataNotDisturb(NotDisturbCollection.GetNotDisturbCollections(cityControlAddNotDisturb.GetCityId(),
+                    addressControlAddNotDisturb.GetAddressId()));
+        }
+        private void CountryControlAddNotDisturb_CountryChanged(int countryId)
+        {
+            ClearDataOnAddNotDisturbTab("CountryControl");
+            cityControlAddNotDisturb.LoadCity(countryId);
+        }
+        private void CityControlAddNotDisturb_CityChanged(int cityId)
+        {
+            ClearDataOnAddNotDisturbTab("CityControl");
+
+            _maskNumber = City.GetMaskNumberbyCityId(cityId);
+
+            addressControlAddNotDisturb.LoadAddress(cityId, chkPrivateHouseAddNotDisturb.Checked);
+
+            chkPrivateHouseAddNotDisturb.Enabled = true;
+            txtNumberPhone.Mask = _maskNumber;
+
+            var notDisturbCollection = NotDisturbCollection.GetNotDisturbCollections(cityId);
+
+            UpdateDataNotDisturb(notDisturbCollection);
+        }
+        private void AddressControlAddNotDisturb_AddressChanged(int addressId)
+        {
+            ClearDataOnAddNotDisturbTab("AddressControl");
+
+            var notDisturbCollection = NotDisturbCollection.GetNotDisturbCollections(cityControlAddNotDisturb.GetCityId(), addressId);
+
+            UpdateDataNotDisturb(notDisturbCollection);
+
+            apartmentControlAddNotDisturb.LoadApartment(addressId);
+        }
+
+        private void ApartmentControlAddNotDisturb_ApartmentChanged(int apartmentId)
+        {
+
+        }
+
+        private void ClearDataOnAddNotDisturbTab(string nameControl)
+        {
+            switch (nameControl)
+            {
+                case "CountryControl":
+                    cityControlAddNotDisturb.ClearDataCity();
+                    chkPrivateHouseAddNotDisturb.Checked = false;
+                    chkPrivateHouseAddNotDisturb.Enabled = false;
+                    goto case "CityControl";
+                case "CityControl":
+                    addressControlAddNotDisturb.ClearDataAddress();
+                    txtNumberPhone.Text = "";
+                    txtNumberPhone.Mask = null;
+                    UpdateDataNotDisturb(new List<NotDisturbCollection>());
+                    goto case "AddressControl";
+                case "AddressControl":
+                    apartmentControlAddNotDisturb.ClearDataApartment();
+                    break;
+            }
+        }
+
+        private void chkPrivateHouseAddNotDisturb_CheckedChanged(object sender, EventArgs e)
+        {
+            if (!string.IsNullOrEmpty(cityControlAddNotDisturb.GetTextCity()))
+            {
+                addressControlAddNotDisturb.ClearDataAddress();
+                cityControlAddNotDisturb.SelectCity(cityControlAddNotDisturb.GetCityId());
+            }
+        }
+        #endregion
+
+        private void btnAddNotDisturbByAddress_Click(object sender, EventArgs e)
+        {
+            if (string.IsNullOrEmpty(countryControlAddNotDisturb.GetTextCountry()) ||
+                string.IsNullOrEmpty(cityControlAddNotDisturb.GetTextCity()) ||
+                string.IsNullOrEmpty(addressControlAddNotDisturb.GetTextAddress()) ||
+                string.IsNullOrEmpty(apartmentControlAddNotDisturb.GetTextApartment()))
+            {
+                MessageBox.Show("Укажите Страну, Город, Адрес и Квартиру.", "Добавить \"Не беспокоить\"",
+                       MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                using (var db = new ApplicationContext())
+                {
+                    var findRecord = db.NotDisturb.Where(n => n.NumberPhoneId == apartmentControlAddNotDisturb.GetApartmentId()).ToList();
+
+                    if (findRecord.Count > 0)
+                    {
+                        MessageBox.Show("Информация об этой квартире уже есть в базе. Проверьте, пожалуйста, введенный адрес.", "Добавить \"Не беспокоить\"",
+                            MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        var addNotDisturb =
+                            new AddNotDisturb($"{addressControlAddNotDisturb.GetTextAddress()}, кв. {apartmentControlAddNotDisturb.GetTextApartment()}");
+                        addNotDisturb.NotDisturb = new NotDisturb() { NumberPhoneId = apartmentControlAddNotDisturb.GetApartmentId() };
+                        if (addNotDisturb.ShowDialog() == DialogResult.OK)
+                        {
+                            var record = addNotDisturb.NotDisturb;
+                            db.NotDisturb.Update(record);
+                            db.SaveChanges();
+
+                            UpdateDataNotDisturb(NotDisturbCollection.GetNotDisturbCollections(cityControlAddNotDisturb.GetCityId(),
+                                addressControlAddNotDisturb.GetAddressId()));
+                        }
+                    }
+                }
+            }
+        }
+
+        private void btnAddNotDisturbByNumberPhone_Click(object sender, EventArgs e)
+        {
+            if (string.IsNullOrEmpty(countryControlAddNotDisturb.GetTextCountry()) ||
+                string.IsNullOrEmpty(cityControlAddNotDisturb.GetTextCity()))
+            {
+                MessageBox.Show("Укажите Страну и Город.", "Добавить \"Не беспокоить\"",
+                       MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                using (var db = new ApplicationContext())
+                {
+                    Regex regex = new Regex(@"[^0-9]");
+                    var findRecord = db.NotDisturb.Include(n => n.NumberPhone).ToList().Where(n => n.NumberPhone?.Number == regex.Replace(txtNumberPhone.Text, "")).ToList();
+                    if (findRecord.Count > 0)
+                    {
+                        MessageBox.Show("Информация об этой квартире уже есть в базе. Проверьте, пожалуйста, введенный адрес.", "Добавить \"Не беспокоить\"",
+                            MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        var numberPhoneId = findRecord[0].NumberPhone?.Id;
+                        var addNotDisturb =
+                            new AddNotDisturb($"{addressControlAddNotDisturb.GetTextAddress()}, кв. {apartmentControlAddNotDisturb.GetTextApartment()}");
+                        addNotDisturb.NotDisturb = new NotDisturb() { NumberPhoneId = apartmentControlAddNotDisturb.GetApartmentId() };
+                        if (addNotDisturb.ShowDialog() == DialogResult.OK)
+                        {
+                            var record = addNotDisturb.NotDisturb;
+                            db.NotDisturb.Update(record);
+                            db.SaveChanges();
+
+                            UpdateDataNotDisturbBetween();
+                        }
+                    }
+
+                }
+            }
+        }
+
+        private void btnEditNotDisturb_Click(object sender, EventArgs e)
+        {
+            if (currentNotDisturb != null)
+            {
+                var editNotDisturb = new AddNotDisturb(addressControlAddNotDisturb.GetTextAddress());
+                using (var db = new ApplicationContext())
+                {
+                    var notDisturb = db.NotDisturb.Where(n => n.Id == currentNotDisturb.Id).FirstOrDefault();
+                    if (notDisturb != null)
+                        editNotDisturb.NotDisturb = notDisturb;
+                }
+                if (editNotDisturb.ShowDialog() == DialogResult.OK)
+                {
+                    using (var db = new ApplicationContext())
+                    {
+                        var singleNumber = editNotDisturb.NotDisturb;
+                        db.NotDisturb.Update(singleNumber);
+                        db.SaveChanges();
+
+                        UpdateDataNotDisturbBetween();
+                    }
+                }
+            }
+            else
+            {
+                MessageBox.Show("Выберите запись для редактирования.", "Добавить \"Не беспокоить\"",
                        MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
         }
